% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rd_files.R
\name{create_rd_files}
\alias{create_rd_files}
\title{Creat References for Elements of Data Dictionary}
\usage{
create_rd_files(dd)
}
\arguments{
\item{dd}{a propoerly formatted data dictionary as tibble. Readin from dd_read}
}
\description{
Creat References for Elements of Data Dictionary
}
