% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_build_site.R
\name{dd_build_site}
\alias{dd_build_site}
\title{Build a Data Dictionary Website for Your Project}
\usage{
dd_build_site(website_path, dd_file)
}
\arguments{
\item{website_path}{Path to where your project's data dictionary website will be stored}

\item{dd_file}{File to  your project's data dictionary}
}
\description{
Build a Data Dictionary Website for Your Project
}
