% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_read.R
\name{dd_read}
\alias{dd_read}
\title{Read in Data Dictionary}
\usage{
dd_read(path_to_dd)
}
\arguments{
\item{path_to_dd}{path to the xls or xlsx formatted data dictionary}
}
\value{

}
\description{
Read in Data Dictionary
}
