% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_read.R
\name{dd_read}
\alias{dd_read}
\title{Read in Data Dictionary}
\usage{
dd_read(dd_file)
}
\arguments{
\item{dd_file}{path to the csv formatted data dictionary}
}
\value{
a tibble.
}
\description{
Read in Data Dictionary
}
